+++ lib/src/cli/flutter_cli.dart
@@ -45,7 +45,7 @@ class FlutterCli {
     bool recursive = false,
   }) async {
     await _runCommand(
-      cmd: (cwd) => _Cmd.run(
+      cmd: (cwd) => _Cmd.start(
         'flutter',
         ['pub', 'run', 'build_runner', 'build', '--delete-conflicting-outputs'],
         workingDirectory: cwd,
@@ -62,8 +62,7 @@ class FlutterCli {
     required Logger logger,
   }) async {
     final check = logger.progress('Checking intl_utils');
-    final res = await _Cmd.run('flutter', ['pub', 'global', 'list'],
-        workingDirectory: cwd);
+    final res = await _Cmd.run('flutter', ['pub', 'global', 'list'], workingDirectory: cwd);
     if (res.stdout.toString().contains('intl_utils')) {
       check.complete('intl_utils is already activated');
       await _generate(logger, cwd);
@@ -78,16 +77,14 @@ class FlutterCli {
 
   static bool isFlutterProject() {
     final currentDirectory = Directory.current;
-    final pubspecFile =
-        File(p.join(currentDirectory.absolute.path, 'pubspec.yaml'));
+    final pubspecFile = File(p.join(currentDirectory.absolute.path, 'pubspec.yaml'));
     return pubspecFile.existsSync();
   }
 
   static String packageName() {
     final currentDirectory = Directory.current;
     if (isFlutterProject()) {
-      final pubspecFile =
-          File(p.join(currentDirectory.absolute.path, 'pubspec.yaml'));
+      final pubspecFile = File(p.join(currentDirectory.absolute.path, 'pubspec.yaml'));
       final yaml = Pubspec.parse(pubspecFile.readAsStringSync());
       return yaml.name;
     } else {
@@ -142,8 +139,7 @@ class FlutterCli {
   }
 
   static Future<void> _generate(Logger logger, String cwd) async {
-    final generate = logger.progress(
-        'Running ${lightGreen.wrap('flutter pub global run intl_utils:generate')}');
+    final generate = logger.progress('Running ${lightGreen.wrap('flutter pub global run intl_utils:generate')}');
     await _Cmd.run(
       'flutter',
       ['pub', 'global', 'run', 'intl_utils:generate'],
@@ -153,22 +149,14 @@ class FlutterCli {
   }
 
   static copyEnvs(Logger logger, String path) async {
-    final envCopy = logger.progress(
-        'Generating .env files for development, staging and production');
-    await _Cmd.run('cp', ['.env.example', '.env-production'],
-        workingDirectory: path);
-    await _Cmd.run('cp', ['.env.example', '.env-development'],
-        workingDirectory: path);
-    await _Cmd.run('cp', ['.env.example', '.env-staging'],
-        workingDirectory: path);
-    envCopy.complete(
-        'Generated .env files for development, staging and production');
+    final envCopy = logger.progress('Generating .env files for development, staging and production');
+    await _Cmd.run('cp', ['.env.example', '.env-production'], workingDirectory: path);
+    await _Cmd.run('cp', ['.env.example', '.env-development'], workingDirectory: path);
+    await _Cmd.run('cp', ['.env.example', '.env-staging'], workingDirectory: path);
+    envCopy.complete('Generated .env files for development, staging and production');
   }
 
-  static Future<void> runProject(
-      {String cwd = '.', String flavor = 'development'}) async {
-    await _Cmd.run(
-        'flutter', ['run', '-t', 'lib/main-$flavor.dart', '--flavor', flavor],
-        workingDirectory: cwd);
+  static Future<void> runProject({String cwd = '.', String flavor = 'development'}) async {
+    await _Cmd.start('flutter', ['run', '-t', 'lib/main-$flavor.dart', '--flavor', flavor], workingDirectory: cwd);
   }
 }