+++ lib/src/commands/generate/subcommands/bloc_command.dart
@@ -34,8 +34,7 @@ class BlocCommand extends Command<int> {
   String get name => 'bloc';
 
   @override
-  String get invocation =>
-      'eb_clean generate bloc --feature <feature-name> <name>';
+  String get invocation => 'eb_clean generate bloc --feature <feature-name> <name>';
 
   @override
   String get summary => '$invocation\n$description';
@@ -43,7 +42,8 @@ class BlocCommand extends Command<int> {
   @override
   Future<int> run() async {
     if (argResults!['feature'] == null) {
-      throw UsageException('feature is required', usage);
+      logger.info('${red.wrap('Feature name is required. please provide feature name with --feature option.')}');
+      return ExitCode.noInput.code;
     }
     final args = argResults?.rest;
     if (args != null && args.isNotEmpty) {
@@ -51,18 +51,14 @@ class BlocCommand extends Command<int> {
       final blocName = args.first;
       final blocTemplate = BlocTemplate();
       String path = '${blocTemplate.path}/$featureName/presentation/blocs/';
-      final blocDone =
-          logger.progress('Generating ${blocName.pascalCase}Bloc class');
-      final blocGenerator =
-          await MasonGenerator.fromBundle(blocTemplate.bundle);
+      final blocDone = logger.progress('Generating ${blocName.pascalCase}Bloc class');
+      final blocGenerator = await MasonGenerator.fromBundle(blocTemplate.bundle);
       var vars = <String, dynamic>{
         'name': blocName,
       };
       final cwd = Directory(p.join(Directory.current.path, path));
-      await blocGenerator.generate(DirectoryGeneratorTarget(cwd),
-          fileConflictResolution: FileConflictResolution.overwrite, vars: vars);
-      blocDone.complete(
-          'Generated ${blocName.pascalCase}Bloc class in ${cwd.path}');
+      await blocGenerator.generate(DirectoryGeneratorTarget(cwd), fileConflictResolution: FileConflictResolution.overwrite, vars: vars);
+      blocDone.complete('Generated ${blocName.pascalCase}Bloc class in ${cwd.path}');
       await blocTemplate.onGenerateComplete(logger, Directory.current);
     } else {
       throw UsageException('please provide bloc name', usage);