+++ lib/src/commands/generate/subcommands/page_command.dart
@@ -22,14 +22,11 @@ class PageCommand extends Command<int> {
         'feature',
         abbr: 'f',
         help: 'feature name to create page',
-        mandatory: true,
       )
-      ..addOption(
-        'type',
-        abbr: 't',
-        help: 'type of page',
-        defaultsTo: 'stateless',
-        allowed: ['stateless', 'stateful'],
+      ..addFlag(
+        'state',
+        abbr: 's',
+        help: 'Creates a stateful page instead of stateless page.',
       );
   }
 
@@ -42,8 +39,7 @@ class PageCommand extends Command<int> {
   String get name => 'page';
 
   @override
-  String get invocation =>
-      'eb_clean generate page --feature <feature-name> --type <stateless,stateful> <name>';
+  String get invocation => 'eb_clean generate page --feature <feature-name> <name>';
 
   @override
   String get summary => '$invocation\n$description';
@@ -51,36 +47,32 @@ class PageCommand extends Command<int> {
   @override
   Future<int> run() async {
     if (argResults!['feature'] == null) {
-      throw UsageException('feature is required', usage);
+      logger.info('${red.wrap('Feature name is required. please provide feature name with --feature option')}');
+      return ExitCode.noInput.code;
     }
 
-    final type = argResults!['type'] as String? ?? 'stateless';
+    final isStateful = argResults!['type'] == true;
     final args = argResults?.rest;
     if (args != null && args.isNotEmpty) {
       final featureName = argResults!['feature'] as String;
       final pageName = args.first;
       final pageTemplate = PageTemplate();
       String path = '${pageTemplate.path}/$featureName/presentation/pages/';
-      final pageDone =
-          logger.progress('Generating ${pageName.pascalCase}Page class');
-      final pageGenerator =
-          await MasonGenerator.fromBundle(pageTemplate.bundle);
+      final pageDone = logger.progress('Generating ${pageName.pascalCase}Page class');
+      final pageGenerator = await MasonGenerator.fromBundle(pageTemplate.bundle);
       var vars = <String, dynamic>{
         'name': pageName,
-        'state': type == 'stateful',
+        'state': isStateful,
         'feature': featureName,
       };
       final cwd = Directory(p.join(Directory.current.path, path));
-      await pageGenerator.generate(DirectoryGeneratorTarget(cwd),
-          fileConflictResolution: FileConflictResolution.overwrite, vars: vars);
+      await pageGenerator.generate(DirectoryGeneratorTarget(cwd), fileConflictResolution: FileConflictResolution.overwrite, vars: vars);
       await pageGenerator.hooks.postGen(
         vars: vars,
         onVarsChanged: (v) => vars = v,
-        workingDirectory:
-            p.join(Directory.current.path, pageTemplate.path, featureName),
+        workingDirectory: p.join(Directory.current.path, pageTemplate.path, featureName),
       );
-      pageDone.complete(
-          'Generated ${pageName.pascalCase}Page class in $featureName feature');
+      pageDone.complete('Generated ${pageName.pascalCase}Page class in $featureName feature');
     } else {
       throw UsageException('please provide bloc name', usage);
     }