+++ lib/src/commands/generate/subcommands/repository_command.dart
@@ -18,8 +18,7 @@ import '../templates/shared/repository/repository.dart';
 class RepositoryCommand extends Command<int> {
   /// {@macro repository_command}
   RepositoryCommand(this.logger) {
-    argParser.addOption('feature',
-        abbr: 'f', help: 'feature name to create repository');
+    argParser.addOption('feature', abbr: 'f', help: 'feature name to create repository');
   }
 
   final Logger logger;
@@ -40,31 +39,27 @@ class RepositoryCommand extends Command<int> {
   Future<int> run() async {
     final packageName = FlutterCli.packageName();
     if (argResults!['feature'] == null) {
-      throw UsageException(
-          'please provide feature to create repository in', usage);
+      logger.info('${red.wrap('Feature name is required. please provide feature name with --feature option.')}');
+      return ExitCode.noInput.code;
     }
     final args = argResults?.rest;
     if (args != null && args.isNotEmpty) {
       final repositoryName = args.first;
       final featureName = argResults!['feature'] as String;
       final repositoryTemplate = RepositoryTemplate();
-      final repositoryDone = logger.progress(
-          'Generating ${repositoryName.pascalCase}Repository\'s abstract and implementation class');
-      final repositoryGenerator =
-          await MasonGenerator.fromBundle(repositoryTemplate.bundle);
+      final repositoryDone = logger.progress('Generating ${repositoryName.pascalCase}Repository\'s abstract and implementation class');
+      final repositoryGenerator = await MasonGenerator.fromBundle(repositoryTemplate.bundle);
       var vars = <String, dynamic>{
         'name': repositoryName,
         'package_name': packageName,
       };
-      final cwd = Directory(p.join(
-          Directory.current.path, repositoryTemplate.path, '$featureName/'));
+      final cwd = Directory(p.join(Directory.current.path, repositoryTemplate.path, '$featureName/'));
       await repositoryGenerator.generate(
         DirectoryGeneratorTarget(cwd),
         fileConflictResolution: FileConflictResolution.overwrite,
         vars: vars,
       );
-      repositoryDone.complete(
-          'Generated ${repositoryName.pascalCase}Repository class in $featureName feature');
+      repositoryDone.complete('Generated ${repositoryName.pascalCase}Repository class in $featureName feature');
       await repositoryTemplate.onGenerateComplete(logger, Directory.current);
     } else {
       throw UsageException('please provide repository name', usage);