+++ lib/src/commands/generate/subcommands/subcommands.dart
@@ -7,6 +7,7 @@
 
 export 'bloc_command.dart';
 export 'cubit_command.dart';
+export 'env_command.dart';
 export 'feature_command.dart';
 export 'page_command.dart';
 export 'repository_command.dart';