+++ lib/src/commands/generate/templates/graphql/source/graphql_source_bundle.dart
@@ -3,46 +3,4 @@
 
 import 'package:mason/mason.dart';
 
-final graphqlSourceBundle = MasonBundle.fromJson(<String, dynamic>{
-  "files": [
-    {
-      "path": "{{name.snakeCase()}}_remote_source.dart",
-      "data":
-          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmluamVjdGFibGUvaW5qZWN0YWJsZS5kYXJ0JzsKCgppbXBvcnQgJ3BhY2thZ2U6Z3JhcGhxbC9jbGllbnQuZGFydCc7CgppbXBvcnQgJ3BhY2thZ2U6e3twYWNrYWdlX25hbWUuc25ha2VDYXNlKCl9fS9zcmMvY29yZS9iYXNlL2Jhc2VfZ3JhcGhfcmVtb3RlX3NvdXJjZS5kYXJ0JzsKCmFic3RyYWN0IGNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSB7Cn0KCkBMYXp5U2luZ2xldG9uKGFzOiB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2UpCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZUltcGwgZXh0ZW5kcyBCYXNlR3JhcGhSZW1vdGVTb3VyY2UgaW1wbGVtZW50cyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2Ugewp7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2VJbXBsKEdyYXBoUUxDbGllbnQgc3VwZXIuY2xpZW50KTsKfQo=",
-      "type": "text"
-    }
-  ],
-  "hooks": [],
-  "name": "graphql_source",
-  "description": "A new brick created with the Mason CLI.",
-  "version": "0.1.0+1",
-  "environment": {"mason": ">=0.1.0-dev <0.1.0"},
-  "readme": {
-    "path": "README.md",
-    "data":
-        "IyBzb3VyY2UKCkEgbmV3IGJyaWNrIGNyZWF0ZWQgd2l0aCB0aGUgTWFzb24gQ0xJLgoKX0dlbmVyYXRlZCBieSBbbWFzb25dWzFdIPCfp7FfCgojIyBHZXR0aW5nIFN0YXJ0ZWQg8J+agAoKVGhpcyBpcyBhIHN0YXJ0aW5nIHBvaW50IGZvciBhIG5ldyBicmljay4KQSBmZXcgcmVzb3VyY2VzIHRvIGdldCB5b3Ugc3RhcnRlZCBpZiB0aGlzIGlzIHlvdXIgZmlyc3QgYnJpY2sgdGVtcGxhdGU6CgotIFtPZmZpY2lhbCBNYXNvbiBEb2N1bWVudGF0aW9uXVsyXQotIFtDb2RlIGdlbmVyYXRpb24gd2l0aCBNYXNvbiBCbG9nXVszXQotIFtWZXJ5IEdvb2QgTGl2ZXN0cmVhbTogRmVsaXggQW5nZWxvdiBEZW1vcyBNYXNvbl1bNF0KClsxXTogaHR0cHM6Ly9naXRodWIuY29tL2ZlbGFuZ2VsL21hc29uClsyXTogaHR0cHM6Ly9naXRodWIuY29tL2ZlbGFuZ2VsL21hc29uL3RyZWUvbWFzdGVyL3BhY2thZ2VzL21hc29uX2NsaSNyZWFkbWUKWzNdOiBodHRwczovL3Zlcnlnb29kLnZlbnR1cmVzL2Jsb2cvY29kZS1nZW5lcmF0aW9uLXdpdGgtbWFzb24KWzRdOiBodHRwczovL3lvdXR1LmJlL0c0UFRqQTZ0cFRVCg==",
-    "type": "text"
-  },
-  "changelog": {
-    "path": "CHANGELOG.md",
-    "data": "IyAwLjEuMCsxCgotIFRPRE86IERlc2NyaWJlIGluaXRpYWwgcmVsZWFzZS4K",
-    "type": "text"
-  },
-  "license": {
-    "path": "LICENSE",
-    "data": "VE9ETzogQWRkIHlvdXIgbGljZW5zZSBoZXJlLgo=",
-    "type": "text"
-  },
-  "vars": {
-    "name": {
-      "type": "string",
-      "description": "Remote Source Name",
-      "prompt": "What is your remote source name?"
-    },
-    "package_name": {
-      "type": "string",
-      "description": "Package Name",
-      "prompt": "What is your remote source name?"
-    }
-  }
-});
+final graphqlSourceBundle = MasonBundle.fromJson(<String, dynamic>{"files":[{"path":"{{name.snakeCase()}}_remote_source.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmluamVjdGFibGUvaW5qZWN0YWJsZS5kYXJ0JzsKCgppbXBvcnQgJ3BhY2thZ2U6Z3JhcGhxbC9jbGllbnQuZGFydCc7CgppbXBvcnQgJ3BhY2thZ2U6e3twYWNrYWdlX25hbWUuc25ha2VDYXNlKCl9fS9zcmMvY29yZS9iYXNlL2Jhc2VfZ3JhcGhfcmVtb3RlX3NvdXJjZS5kYXJ0JzsKCmFic3RyYWN0IGNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSB7Cn0KCkBMYXp5U2luZ2xldG9uKGFzOiB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2UpCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZUltcGwgZXh0ZW5kcyBCYXNlR3JhcGhSZW1vdGVTb3VyY2UgaW1wbGVtZW50cyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2Ugewp7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2VJbXBsKEdyYXBoUUxDbGllbnQgc3VwZXIuY2xpZW50KTsKfQo=","type":"text"}],"hooks":[],"name":"graphql_source","description":"A new brick created with the Mason CLI.","version":"0.1.0+1","environment":{"mason":">=0.1.0-dev <0.1.0"},"readme":{"path":"README.md","data":"IyBzb3VyY2UKCkEgbmV3IGJyaWNrIGNyZWF0ZWQgd2l0aCB0aGUgTWFzb24gQ0xJLgoKX0dlbmVyYXRlZCBieSBbbWFzb25dWzFdIPCfp7FfCgojIyBHZXR0aW5nIFN0YXJ0ZWQg8J+agAoKVGhpcyBpcyBhIHN0YXJ0aW5nIHBvaW50IGZvciBhIG5ldyBicmljay4KQSBmZXcgcmVzb3VyY2VzIHRvIGdldCB5b3Ugc3RhcnRlZCBpZiB0aGlzIGlzIHlvdXIgZmlyc3QgYnJpY2sgdGVtcGxhdGU6CgotIFtPZmZpY2lhbCBNYXNvbiBEb2N1bWVudGF0aW9uXVsyXQotIFtDb2RlIGdlbmVyYXRpb24gd2l0aCBNYXNvbiBCbG9nXVszXQotIFtWZXJ5IEdvb2QgTGl2ZXN0cmVhbTogRmVsaXggQW5nZWxvdiBEZW1vcyBNYXNvbl1bNF0KClsxXTogaHR0cHM6Ly9naXRodWIuY29tL2ZlbGFuZ2VsL21hc29uClsyXTogaHR0cHM6Ly9naXRodWIuY29tL2ZlbGFuZ2VsL21hc29uL3RyZWUvbWFzdGVyL3BhY2thZ2VzL21hc29uX2NsaSNyZWFkbWUKWzNdOiBodHRwczovL3Zlcnlnb29kLnZlbnR1cmVzL2Jsb2cvY29kZS1nZW5lcmF0aW9uLXdpdGgtbWFzb24KWzRdOiBodHRwczovL3lvdXR1LmJlL0c0UFRqQTZ0cFRVCg==","type":"text"},"changelog":{"path":"CHANGELOG.md","data":"IyAwLjEuMCsxCgotIFRPRE86IERlc2NyaWJlIGluaXRpYWwgcmVsZWFzZS4K","type":"text"},"license":{"path":"LICENSE","data":"VE9ETzogQWRkIHlvdXIgbGljZW5zZSBoZXJlLgo=","type":"text"},"vars":{"name":{"type":"string","description":"Remote Source Name","prompt":"What is your remote source name?"},"package_name":{"type":"string","description":"Package Name","prompt":"What is your remote source name?"}}});