+++ lib/src/commands/packages_command/subcommands/build_runner_command.dart
@@ -21,11 +21,11 @@ class BuildRunnerCommand extends Command<int> {
       negatable: false,
     );
   }
+
   final Logger logger;
 
   @override
-  String get description =>
-      'Runs flutter pub run build_runner build --delete-conflicting-outputs in current directory';
+  String get description => 'Runs flutter pub run build_runner build --delete-conflicting-outputs in current directory';
 
   @override
   String get name => 'build_runner';
@@ -40,16 +40,11 @@ class BuildRunnerCommand extends Command<int> {
   Future<int> run() async {
     if (FlutterCli.isFlutterProject()) {
       final recursive = argResults!['recursive'] == true;
-      final pubDone = logger.progress(
-          'Running ${lightGreen.wrap('flutter pub run build_runner build --delete-conflicting-outputs')}');
-      await FlutterCli.runBuildRunner(
-          cwd: Directory.current.path, recursive: recursive);
-      pubDone.complete();
+      logger.info('Running ${lightGreen.wrap('flutter pub run build_runner build --delete-conflicting-outputs')}');
+      await FlutterCli.runBuildRunner(cwd: Directory.current.path, recursive: recursive);
       return ExitCode.success.code;
     } else {
-      throw UsageException(
-          'packages build_runner only available inside flutter project only',
-          usage);
+      throw UsageException('packages build_runner only available inside flutter project only', usage);
     }
   }
 }